bin: Print an apology requesting a bug report on panic
As of this commit, if `cargo run -p holochain_2020` panics,
it prints:

```none
Well, this is embarrassing.

holochain_2020 had a problem and crashed. To help us diagnose the problem you can send us a crash report.

We have generated a report file at "/tmp/report-574136e7-698a-4f4d-a8de-223f9739dcdb.toml". Submit an issue or email with the subject of "holochain_2020 Crash Report" and include the report as an attachment.

- Homepage: https://github.com/Holo-host/holochain-2020
- Authors: Holochain Core Dev Team <devcore@holochain.org>

We take privacy seriously, and do not perform any automated error collection. In order to improve the software, we rely on people to submit reports.

Thank you kindly!
```