WIP: update tokio to 0.3
This updates the tokio dependency in all crates to tokio 0.3.

The update approach is as mechanical as possible, i.e. it's trying to
purely adapt to the changes introduced in tokio 0.3 without unrelated
refactors.

These out-of-repo dependencies were updated to tokio 0.3 in the process
as well:
tokio 0.3 compatible versions:
* tokio_safe_block_on:
* lair

A new crate `tokio_helper` was introduced to work around issues with the
`updated tokio_safe_block_on`. It provies functionality to use a
globally and statically available tokio runtime to be entered from
anywhere blocking calls are required.

Currently this work is blocked on two dependencies that haven't updated
to tokio 0.3 yet:
* quinn: used by the `transoprt_quic` crate
* reqwest