test_utils/wasm/build.rs: make 'cargo check' opt-out
This introduces a feature named `only_check_wasms` for the case where we
want to check but not build the crate at
`crates/test_utils/wasm/wasm_workspace`.

The new feature is enabled by default but can be disabled by passing
`--no-default-features` to `cargo` which will then effectively avoid any
operation on the `wasm_workspace`crate. This saves effort on non-test
related builds.