ci(release): use additional space on github actions
after we've got the extended space mounted, we can use overlayfs and
bind mounts as appropriate to make this space available to all paths
that are expected to be written to.

for new directories (such as /nix), a bind mount is enough for existing
directories (such as $HOME and /tmp), we use an overlay to preserve
access to the existing files, while redirecting changes to the extended
space.