WIP add `get_local_authors` host call to fix sharded authorship problem
However, we need to figure out how to fully search the holochain
database for a particular kitsune basis, since the latter has no context
other than the raw hash, and in holochain, we have several types of
basis with different representations in the database.

Either we need to represent all bases using the raw hash, or we need to
search every possible basis representation for a given kitsune basis.