debugging websocket errors
the way use websockets in the shared values remote impl doesn't work
beyond 1 message. if the sender is reused to send another message it
throws a `FailedToRecvResp` error. when i use `connect` for every
message it works, but this seems incredibly inefficient and i hope this
is a misunderstanding on my side rather than a limitation of
`holochain_websocket`.