test: remove out-dated test
Originally, it was to test that an identical op is only fetched from the
network once and then looked up in the cache. After a refactor of
production code this was no longer the case and so the test was
refactored to check that it can fetch from the network multiple times.
There can be no guarantee that it will do one over the other so the test
is naturally flaky.