fix: Remove `same_dht` field from `SysValDeps` (#5243)
* fix: Remove `same_dht` field from `SysValDeps`

The field is speculative in case we add the ability to fetch ops from other DHTs when validating but there is no plan to implement that and the type is internal. For now, it's just more effort to read than simply getting the current validation dependencies which can only come from the current DHT anyway.

* docs: Update changelog